Add the framework for MessageHandler
The MessageHandler is meant to be used by SMTP and web ui code that send
messages, to provide a higher level interface to replace Channel, while
handling complications such as resending messages across several channels due
to timeouts. This commit is the start of a series of commits that moves
code and reponsibilities from other classes (mostly Channel) to
MessageHandler in order to make the channel timeouts nearly invisible to
the rest of the code.