Create a new connection on each WoT request
At the moment this seems to be the only reliable way to detect that a
plugin has been unloaded since PluginTalker.send() doesn't throw if the
plugin has been unloaded. The overhead of doing this seems to be no more
than a few extra object creations per request. As a side effect of this
multiple WoT requests can be processed concurrently and it is possible
to safely add timeouts to the requests.