Only allow base32 match when using SMTP
This is needed because the recipient received via SMTP should only be
checked using the base32 test because the domain name part of an email
address is case insensitive. Because of that a 'helpful' email client
could change the case which could make the base64 test match the wrong
identity (with partial matches this is not extremely unlikely).