Add timeout to message fetching
This timeout means that receiving a lot of messages from one contact
won't block the other operations done in SingleAccountWatcher. In most
cases this means that we will fetch one message then return since
fetches are slow. This will reduce the overall throughput since we check
for RTS messages each round in SingleAccountWatcher.