Add timeout to ack fetching
This adds a timeout to the ack fetching to avoid stopping other tasks if
there are many messages waiting for acks. To avoid checking the same
acks over and over while not checking others we store the index of the
next ack that should be checked. This means that we might skip some acks
when messages are removed from the outbox, but this should not be a
problem since we will eventually wrap around.