Add test for literal without linebreak (bug 5350)
The test fails at the moment since the bug hasn't been fixed. At the
moment the failure is ignored if the expected and actual values don't
change, and a line is printed instead. This was the best solution I
could think of that would allow the test to be run as usual without
disabling checking for any other tests.