Fix race condition when receiving duplicate messages
There was a check-then-act race between the calls to the isPresent() and
add() methods of the message log. This could cause messages to be
duplicated if two versions (e.g. original and resent) were fetched at
approximately the same time. There is a minor regression in this patch
since the outbox creation won't be repeated if it fails, and won't be
attempted if it is deleted while Freemail is running.