Rewrite parts of HighLevelFCPClient to be parallel
The FCP messages have a unique id making it possible to match each reply
to the request. Instead of keeping one reply a list is now used and each
thread will look for the correct identifier in the reply list. There is
a race condition when a thread is interrupted since the cancel might
happen after the FCPConnection has called requestFinished(), however it
will only cause a slight memory leak, and interrupting the thread only
happens when shutting down.