checkcommit: don't check out the original commit
checkcommit depended on the output of git branch which it should never
have done since it isn't stable. Looking at the git-prompt.sh code from
git-contrib, getting back to the original state in the way we wanted
would be really hard, so just skip it. checkcommit was always supposed
to be a low-level abstraction of the build process anyway.