Mock Executor to fix ordering during transport test
Due to the way operations happen when sending a message a five minute delay
usually occurs when the RTS keys need to be generated. This change mocks the
executor that runs the tasks, and ignores the retry delay when running the
message handler test so it will complete in a reasonable amount of time.