Automatically disconnect after 30 minutes
I have a couple of Freemail accounts, and Freemail over the time of days
starts collecting more and more threads, all hanging in the readLine()
method in the IMAPHandler. Stepping through the code in a debugger
showed that the VM thought that all those sockets it was reading from
were still connected. However, both operating systems (the one my node
runs on and the one my mail client runs on) were in agreement that the
actual number of connections was way smaller.

Now an IMAP handler thread disconnects after a 30 minute timeout, and
the problem has not reappeared since.